Repository is now a ros2 package, to be built with colcon and callable from ros2 cli
