+++ package.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>fleet_adapter_mir</name>
+  <version>0.1.0</version>
+  <description>RMF fleet adapter for MiR robots</description>
+  <maintainer email="grey@openrobotics.org">Grey</maintainer>
+  <maintainer email="aaron@openrobotics.org">Aaron</maintainer>
+  <license>Apache License 2.0</license>
+
+  <exec_depend>rmf_fleet_adapter_python</exec_depend>
+
+  <export>
+    <build_type>ament_python</build_type>
+  </export>
+</package>
