+++ setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/fleet_adapter_mir
+[install]
+install_scripts=$base/lib/fleet_adapter_mir
